Fixed non-AL stuff from Chris
git-svn-id: https://openmw.svn.sourceforge.net/svnroot/openmw/trunk@19 ea6a568a-9f4f-0410-981a-c910a81bb256

+++ bsa/bsafile.d
@@ -177,7 +177,7 @@ class BSAFile
        */
       assert(!isLoaded);
 
-      uint fsize = mmf.length;
+      ulong fsize = mmf.length;
 
       if( fsize < 12 )
 	fail("File too small to be a valid BSA archive");
